Changed line width to be double, fixed width retrieval from symbol, proportional scaling of symbol layers when settings size/width.
git-svn-id: http://svn.osgeo.org/qgis/branches/symbology-ng-branch@11064 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/symbology-ng/qgslinesymbollayerv2.cpp
@@ -8,7 +8,7 @@
 
 #include <cmath>
 
-QgsSimpleLineSymbolLayerV2::QgsSimpleLineSymbolLayerV2(QColor color, int width, Qt::PenStyle penStyle)
+QgsSimpleLineSymbolLayerV2::QgsSimpleLineSymbolLayerV2(QColor color, double width, Qt::PenStyle penStyle)
  : mPenStyle(penStyle), mOffset(0)
 {
   mColor = color;
@@ -19,13 +19,13 @@ QgsSimpleLineSymbolLayerV2::QgsSimpleLineSymbolLayerV2(QColor color, int width, 
 QgsSymbolLayerV2* QgsSimpleLineSymbolLayerV2::create(const QgsStringMap& props)
 {
   QColor color = DEFAULT_SIMPLELINE_COLOR;
-  int width = DEFAULT_SIMPLELINE_WIDTH;
+  double width = DEFAULT_SIMPLELINE_WIDTH;
   Qt::PenStyle penStyle = DEFAULT_SIMPLELINE_PENSTYLE;
   
   if (props.contains("color"))
     color = QgsSymbolLayerV2Utils::decodeColor(props["color"]);
   if (props.contains("width"))
-    width = props["width"].toInt();
+    width = props["width"].toDouble();
   if (props.contains("penstyle"))
     penStyle = QgsSymbolLayerV2Utils::decodePenStyle(props["penstyle"]);
